Create rich text editor using mantine ui
The global stylesheet lacks ol and ul list styles which makes them
invisible. By setting it explicitly, we can fix the issue.

Since the rich text editor return a HTML string, we can directly
render the HTML. To ensure safety and prevent XSS or injection
attacks, we use DOMPurify to sanitize the input.